Refactor data-tables dir to issue-tables
